requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
jsonpickle==2.0.0
gunicorn==20.1.0
django_heroku==0.3.1
